# api/openapi.yaml
openapi: 3.1.0
info:
  version: 0.0.1
  title: Fpan REST API
  license:
    name: MIT
    identifier: MIT
servers:
  - url: "https://pan.hrtius.com/api/v1"
...
security:
  - CookieSession: []
paths:
...
  /auth/login:
    get:
      summary: Start OIDC login
      operationId: startOIDCLogin
      security: []
      responses:
        "302":
          description: Redirect to the OIDC provider
        "500":
          $ref: "#/components/responses/Error"
  /auth/callback:
    get:
      summary: Complete OIDC login
      operationId: completeOIDCLogin
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect after creating the session
        "400":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
...
  /entries:
    get:
      summary: List root entries
      description: Returns files and folders at the root level (parent_id is null).
      operationId: listRootEntries
      tags:
        - Entries
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/SortBy"
        - $ref: "#/components/parameters/Filter"
        - $ref: "#/components/parameters/TypeFilter"
      responses:
        "200":
          $ref: "#/components/responses/PaginatedEntries"
        "400":
          $ref: "#/components/responses/Error"
  /folders/{id}/entries:
    get:
      summary: List folder entries
      description: Returns files and sub-folders inside the specified folder.
      operationId: listFolderEntries
      tags:
        - Entries
      parameters:
        - $ref: "#/components/parameters/FolderId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/SortBy"
        - $ref: "#/components/parameters/Filter"
        - $ref: "#/components/parameters/TypeFilter"
      responses:
        "200":
          $ref: "#/components/responses/PaginatedEntries"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"

  # ============================================================
  #  Files CRUD
  # ============================================================
  /files:
    post:
      summary: Upload file to root (form-data)
      operationId: uploadFileToRoot
      tags:
        - Files
      requestBody:
        $ref: "#/components/requestBodies/UploadFile"
      responses:
        "201":
          $ref: "#/components/responses/FileDetail"
        "400":
          $ref: "#/components/responses/Error"
  /files/stream:
    post:
      summary: Upload file to root (octet-stream)
      operationId: uploadFileToRootStream
      tags:
        - Files
      parameters:
        - $ref: "#/components/parameters/FileDisplay"
        - $ref: "#/components/parameters/FileMimeType"
      requestBody:
        $ref: "#/components/requestBodies/UploadFileStream"
      responses:
        "201":
          $ref: "#/components/responses/FileDetail"
        "400":
          $ref: "#/components/responses/Error"
  /folders/{id}/files:
    post:
      summary: Upload file to folder (form-data)
      operationId: uploadFileToFolder
      tags:
        - Files
      parameters:
        - $ref: "#/components/parameters/FolderId"
      requestBody:
        $ref: "#/components/requestBodies/UploadFile"
      responses:
        "201":
          $ref: "#/components/responses/FileDetail"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /folders/{id}/files/stream:
    post:
      summary: Upload file to folder (octet-stream)
      operationId: uploadFileToFolderStream
      tags:
        - Files
      parameters:
        - $ref: "#/components/parameters/FolderId"
        - $ref: "#/components/parameters/FileDisplay"
        - $ref: "#/components/parameters/FileMimeType"
      requestBody:
        $ref: "#/components/requestBodies/UploadFileStream"
      responses:
        "201":
          $ref: "#/components/responses/FileDetail"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /files/{id}:
    get:
      summary: Get file metadata
      operationId: getFile
      tags:
        - Files
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          $ref: "#/components/responses/FileDetail"
        "404":
          $ref: "#/components/responses/Error"
    put:
      summary: Update file (rename / move)
      operationId: updateFile
      tags:
        - Files
      parameters:
        - $ref: "#/components/parameters/FileId"
      requestBody:
        $ref: "#/components/requestBodies/UpdateFile"
      responses:
        "200":
          $ref: "#/components/responses/FileDetail"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      summary: Delete file (soft)
      operationId: deleteFile
      tags:
        - Files
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "204":
          description: Deleted successfully
        "404":
          $ref: "#/components/responses/Error"

  # ============================================================
  #  Folders CRUD
  # ============================================================
...
  /blobs/{sha256}:
    get:
      summary: Download blob content
      description: Returns the raw binary content of a physical file identified by sha256.
      operationId: getBlobContent
      tags:
        - Blobs
      parameters:
        - $ref: "#/components/parameters/Sha256"
      responses:
        "200":
          description: Blob binary content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/Error"

  # ============================================================
  #  Shares — link management (authenticated)
  # ============================================================
  /shares:
    post:
      summary: Create share link
      operationId: createShare
      tags:
        - Shares
      requestBody:
        $ref: "#/components/requestBodies/CreateShare"
      responses:
        "201":
          $ref: "#/components/responses/ShareDetail"
        "400":
          $ref: "#/components/responses/Error"
    get:
      summary: List my shares
      operationId: listShares
      tags:
        - Shares
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          $ref: "#/components/responses/PaginatedShares"
        "400":
          $ref: "#/components/responses/Error"
  /shares/{id}:
    get:
      summary: Get share detail
      operationId: getShare
      tags:
        - Shares
      parameters:
        - $ref: "#/components/parameters/ShareId"
      responses:
        "200":
          $ref: "#/components/responses/ShareDetail"
        "404":
          $ref: "#/components/responses/Error"
    put:
      summary: Update share
      operationId: updateShare
      tags:
        - Shares
      parameters:
        - $ref: "#/components/parameters/ShareId"
      requestBody:
        $ref: "#/components/requestBodies/UpdateShare"
      responses:
        "200":
          $ref: "#/components/responses/ShareDetail"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      summary: Delete share
      operationId: deleteShare
      tags:
        - Shares
      parameters:
        - $ref: "#/components/parameters/ShareId"
      responses:
        "204":
          description: Deleted successfully
        "404":
          $ref: "#/components/responses/Error"

  # ============================================================
  #  Shared access — public (no auth)
  # ============================================================
  /s/{token}:
    get:
      summary: Access shared resource
      description: |
        Returns metadata for a shared file or folder.
        For password-protected shares, supply `?password=...`.
      operationId: accessShared
      tags:
        - Shared
      security: []
      parameters:
        - $ref: "#/components/parameters/ShareToken"
        - $ref: "#/components/parameters/SharePassword"
      responses:
        "200":
          $ref: "#/components/responses/SharedResource"
        "403":
          $ref: "#/components/responses/Error"
          description: Wrong password, expired, or download limit reached
        "404":
          $ref: "#/components/responses/Error"
  /s/{token}/entries:
    get:
      summary: List shared folder contents
      operationId: listSharedEntries
      tags:
        - Shared
      security: []
      parameters:
        - $ref: "#/components/parameters/ShareToken"
        - $ref: "#/components/parameters/SharePassword"
        - $ref: "#/components/parameters/SharedParentId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/SortBy"
        - $ref: "#/components/parameters/Filter"
      responses:
        "200":
          $ref: "#/components/responses/PaginatedEntries"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /s/{token}/blobs/{sha256}:
    get:
      summary: Download shared file content
      operationId: downloadSharedBlob
      tags:
        - Shared
      security: []
      parameters:
        - $ref: "#/components/parameters/ShareToken"
        - $ref: "#/components/parameters/Sha256"
        - $ref: "#/components/parameters/SharePassword"
      responses:
        "200":
          description: File binary content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    CookieSession:
      description: HttpOnly session cookie created after OIDC login
      type: apiKey
      in: cookie
      name: fpan_session
    OpenIdConnect:
      type: openIdConnect
      openIdConnectUrl: https://pi.auth.hrtius.com/.well-known/openid-configuration

  # ----------------------------------------------------------
  #  Shared Parameters
  # ----------------------------------------------------------
  parameters:
...
    FileId:
      name: id
      in: path
      required: true
      description: File ID
      schema:
        type: integer
        format: int64
    FolderId:
      name: id
      in: path
      required: true
      description: Folder ID
      schema:
        type: integer
        format: int64
...
    Sha256:
      name: sha256
      in: path
      required: true
      description: SHA-256 hex digest (64 lowercase hex chars)
      schema:
        type: string
        minLength: 64
        maxLength: 64
        pattern: "^[a-f0-9]{64}$"
...
    FileDisplay:
      name: X-File-Name
      in: header
      required: true
      description: Original file name
      schema:
        type: string
        minLength: 1
    FileMimeType:
      name: X-File-Type
      in: header
      required: false
      description: File MIME type hint (e.g. image/png)
      schema:
        type: string

    # --- query parameters (shared) ---
...
    Page:
      name: page
      in: query
      required: false
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      required: false
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      required: false
      description: Sort direction
      schema:
        type: string
        enum: [asc, desc]
        default: asc
    SortBy:
      name: sort_by
      in: query
      required: false
      description: Field to sort by
      schema:
        type: string
        enum: [name, created_at, updated_at]
        default: name
    Filter:
      name: filter
      in: query
      required: false
      description: Keyword filter on display name
      schema:
        type: string
    TypeFilter:
      name: type
      in: query
      required: false
      description: Filter by entry type
      schema:
        type: string
        enum: [all, file, folder]
        default: all

  # ----------------------------------------------------------
  #  Schemas
  # ----------------------------------------------------------
...
  requestBodies:
    UploadFile:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              file:
                type: string
                format: binary
                description: The file to upload
            required:
              - file
    UploadFileStream:
      required: true
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary
            description: Raw file bytes
    UpdateFile:
      content:
        application/json:
          schema:
            type: object
            properties:
              display:
                type: string
                minLength: 1
                description: New file name
              parent_id:
                type: [integer, "null"]
                format: int64
                description: Move to another folder (null = root)
...
  responses:
    Error:
      description: Operation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    FileDetail:
      description: File metadata
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/File"
...
    ShareDetail:
      description: Share metadata
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Share"
    PaginatedEntries:
      description: Paginated list of files and folders
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PaginatedEntries"
...